Changed byte array type and improved font IDs
Now, all byte arrays use the type uint8_t instead of unsigned char. Also, the font system was redesigned so that font IDs are actually pointers to the font data rather than an arbitrary enumeration. This allows the compiler to automatically choose what fonts to include instead of the user having to specify it manually at compile time.